requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
boto3>=1.34.129
requests-oauthlib>=2.0.0
cryptography>=42.0.8
python-dotenv>=1.0.1
pymongo==4.5.0
pydantic>=2.6.4
email-validator>=2.2.0
pyjwt>=2.10.1
passlib>=1.7.4
tzdata>=2024.2
motor==3.3.1
pytest>=8.0.0
black>=24.1.1
isort>=5.13.2
flake8>=7.0.0
mypy>=1.8.0
python-jose>=3.3.0
requests>=2.31.0
pandas>=2.2.0
numpy>=2.0.0
scipy>=1.13.0
python-multipart>=0.0.9
jq>=1.6.0
typer>=0.9.0
aiofiles>=23.2.1
httpx>=0.27.0
# emergentintegrations>=0.1.0
sendgrid>=6.10.0

# Phase 4A: Tokuro AI Report Generation Engine
reportlab>=4.0.0  # PDF generation
python-docx>=1.1.0  # DOCX generation
openpyxl>=3.1.0  # XLSX generation
python-pptx>=0.6.23  # PPTX generation
Pillow>=10.0.0  # Image processing for reports
jinja2>=3.1.0  # Template engine for report generation
PyPDF2==3.0.1  # PDF manipulation and reading

# Phase 1: Assessment Framework Setup & Library
PyYAML>=6.0  # YAML file processing
xlrd>=2.0.1  # Excel file reading (legacy support)
chardet>=5.2.0  # Character encoding detection
python-magic>=0.4.27  # File type detection

# Phase 2A: Assessment Engine Core
json-logic>=0.6.3  # JSON Logic for conditional form logic

# MVP 1: Enhanced Authentication & Security
bcrypt>=4.1.0  # Password hashing
python-jose[cryptography]>=3.3.0  # Enhanced JWT with cryptography support
passlib[bcrypt]>=1.7.4  # Password context with bcrypt support

# MVP 1 Week 3: File Upload and Processing
aiofiles>=23.2.1  # Async file operations
python-multipart>=0.0.9  # File upload support

# Tokuro AI Dependencies - DAAKYI's Proprietary AI Engine
openai==1.7.2
filetype==1.2.0
scikit-learn>=1.5.0 # Machine learning library for Tokuro AI

 